move env reader variables to a util mod
