+++ src/main.rs
@@ -11,18 +11,14 @@ pub mod db_connection;
 pub mod handlers;
 pub mod models;
 pub mod schema;
+pub mod util;
 
 use actix_web::{App, HttpServer, web};
 use db_connection::{establish_connection};
+use util::{get_worker_num};
 
 fn main() {
     dotenv::dotenv().unwrap_or_default();
-
-    let worker_key = "LITTLE_LOOKUP_WORKER_NUM";
-    let worker_num = match std::env::var(worker_key) {
-        Ok(val) => val.parse::<usize>().unwrap(),
-        Err(_) => 2
-    };
 
     HttpServer::new(|| {
         App::new()
@@ -48,7 +44,7 @@ fn main() {
                     .route(web::get().to_async(handlers::items::list_items))
             )
     })
-    .workers(worker_num)
+    .workers(get_worker_num())
     .bind("0.0.0.0:8088")
     .unwrap()
     .run()
